Debug Helper: Added WP Error object handling. (#17247)
In case of an error that contains object data, like in case of a disconnection, on shutdown this filter receives a WP_Error. This additional condition helps avoid a fatal error when trying to resolve an object as an array.

Committed via a GitHub action: https://github.com/automattic/jetpack/runs/271375541