PHPCS: Fix most auto-fixable sniffs coming in new WPCS snapshot (#28176)
We're going to update the WordPress-Coding-Standards library from a
2022-02-07 snapshot to a 2023-01-04 snapshot, which brings in a number
of new sniffs. So first we need to fix the stuff the new sniffs will
complain about.

This PR updates for the following sniffs:

* Generic.WhiteSpace.IncrementDecrementSpacing.SpaceAfterIncrement
* Modernize.FunctionCalls.Dirname
* PSR12.Functions.ReturnTypeDeclaration.SpaceBeforeColon
* PSR2.Methods.FunctionClosingBrace.SpacingBeforeClose
* Universal.Constants.UppercaseMagicConstants
* Universal.Operators.DisallowStandalonePostIncrementDecrement

Then it has to do a bit more cleanup on touched lines in some of the
excluded files. Yay tech debt.

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/3841175056